+++ core/src/components/item-sliding/test/shapes/item-sliding.e2e.ts
@@ -7,8 +7,11 @@ import { configs, test, dragElementBy } from '@utils/test/playwright';
  */
 configs({ modes: ['ionic-md'], directions: ['ltr'] }).forEach(({ title, screenshot, config }) => {
   test.describe(title('item-sliding: shapes'), () => {
-    test('should not have visual regressions when not expanded', async ({ page }) => {
+    test('should not have visual regressions when not expanded', async ({ page, skip }) => {
       await page.goto(`/src/components/item-sliding/test/shapes`, config);
+
+      // TODO(FW-7288): Remove skip once fix has been implemented
+      skip.browser('webkit', 'Flaky test in Safari');
 
       const itemIDs = ['round', 'soft', 'rectangular'];
       for (const itemID of itemIDs) {