+++ core/src/components/textarea/test/textarea.spec.tsx
@@ -7,7 +7,8 @@ it('should render bottom content when helper text is defined', async () => {
     html: `<ion-textarea label="Textarea" helper-text="Helper Text"></ion-textarea>`,
   });
 
-  const bottomContent = page.body.querySelector('ion-textarea .textarea-bottom');
+  const textarea = page.body.querySelector('ion-textarea')!;
+  const bottomContent = textarea.shadowRoot!.querySelector('.textarea-bottom');
   expect(bottomContent).not.toBe(null);
 });
 
@@ -17,7 +18,8 @@ it('should render bottom content when helper text is undefined', async () => {
     html: `<ion-textarea label="Textarea"></ion-textarea>`,
   });
 
-  const bottomContent = page.body.querySelector('ion-textarea .textarea-bottom');
+  const textarea = page.body.querySelector('ion-textarea')!;
+  const bottomContent = textarea.shadowRoot!.querySelector('.textarea-bottom');
   expect(bottomContent).toBe(null);
 });
 
@@ -27,6 +29,7 @@ it('should render bottom content when helper text is empty string', async () => 
     html: `<ion-textarea label="Textarea" helper-text=""></ion-textarea>`,
   });
 
-  const bottomContent = page.body.querySelector('ion-textarea .textarea-bottom');
+  const textarea = page.body.querySelector('ion-textarea')!;
+  const bottomContent = textarea.shadowRoot!.querySelector('.textarea-bottom');
   expect(bottomContent).toBe(null);
 });