+++ packages/vue/src/proxies.ts
@@ -819,11 +819,15 @@ export const IonRefresher: StencilVueComponent<JSX.IonRefresher> = /*@__PURE__*/
   'disabled',
   'ionRefresh',
   'ionPull',
-  'ionStart'
+  'ionStart',
+  'ionPullStart',
+  'ionPullEnd'
 ], [
   'ionRefresh',
   'ionPull',
-  'ionStart'
+  'ionStart',
+  'ionPullStart',
+  'ionPullEnd'
 ]);
 
 
@@ -937,6 +941,7 @@ export const IonSegmentContent: StencilVueComponent<JSX.IonSegmentContent> = /*@
 
 export const IonSegmentView: StencilVueComponent<JSX.IonSegmentView> = /*@__PURE__*/ defineContainer<JSX.IonSegmentView>('ion-segment-view', defineIonSegmentView, [
   'disabled',
+  'swipeGesture',
   'ionSegmentViewScroll'
 ], [
   'ionSegmentViewScroll'
@@ -945,6 +950,7 @@ export const IonSegmentView: StencilVueComponent<JSX.IonSegmentView> = /*@__PURE
 
 export const IonSelect: StencilVueComponent<JSX.IonSelect, JSX.IonSelect["value"]> = /*@__PURE__*/ defineContainer<JSX.IonSelect, JSX.IonSelect["value"]>('ion-select', defineIonSelect, [
   'cancelText',
+  'cancelIcon',
   'color',
   'compareWith',
   'disabled',
@@ -986,6 +992,8 @@ export const IonSelect: StencilVueComponent<JSX.IonSelect, JSX.IonSelect["value"
 
 export const IonSelectModal: StencilVueComponent<JSX.IonSelectModal> = /*@__PURE__*/ defineContainer<JSX.IonSelectModal>('ion-select-modal', defineIonSelectModal, [
   'header',
+  'cancelText',
+  'cancelIcon',
   'multiple',
   'options'
 ]);
@@ -993,7 +1001,8 @@ export const IonSelectModal: StencilVueComponent<JSX.IonSelectModal> = /*@__PURE
 
 export const IonSelectOption: StencilVueComponent<JSX.IonSelectOption> = /*@__PURE__*/ defineContainer<JSX.IonSelectOption>('ion-select-option', defineIonSelectOption, [
   'disabled',
-  'value'
+  'value',
+  'description'
 ]);
 
 
@@ -1117,6 +1126,7 @@ export const IonToggle: StencilVueComponent<JSX.IonToggle, JSX.IonToggle["checke
 
 
 export const IonToolbar: StencilVueComponent<JSX.IonToolbar> = /*@__PURE__*/ defineContainer<JSX.IonToolbar>('ion-toolbar', defineIonToolbar, [
-  'color'
+  'color',
+  'titlePlacement'
 ]);