+++ packages/vue/test/base/src/views/Home.vue
@@ -50,6 +50,9 @@
         <ion-item router-link="/tabs-basic" id="tab-basic">
           <ion-label>Tabs with Basic Navigation</ion-label>
         </ion-item>
+        <ion-item router-link="/tabs-similar-prefixes" id="tabs-similar-prefixes">
+          <ion-label>Tabs with Similar Route Prefixes</ion-label>
+        </ion-item>
         <ion-item router-link="/lifecycle" id="lifecycle">
           <ion-label>Lifecycle</ion-label>
         </ion-item>